[Fix] `WatcherInstance` ready event

+++ test/code/platform/watcher.test.ts
@@ -1,9 +1,9 @@
 import * as assert from 'assert';
-import { after, before } from 'mocha';
-import { IDisposable, disposeAll } from 'src/base/common/dispose';
+import { afterEach, before } from 'mocha';
 import { DataBuffer } from 'src/base/common/files/buffer';
 import { Schemas, URI } from 'src/base/common/files/uri';
-import { Blocker } from 'src/base/common/utilities/async';
+import { EventBlocker } from 'src/base/common/utilities/async';
+import { directoryExists } from 'src/base/node/io';
 import { FileService, IFileService } from 'src/platform/files/common/fileService';
 import { IRawResourceChangeEvents } from 'src/platform/files/common/watcher';
 import { DiskFileSystemProvider } from 'src/platform/files/node/diskFileSystemProvider';
@@ -15,32 +15,34 @@ suite('watcher-test', () => {
     let fileService: IFileService;
     const baseURI = URI.join(TestURI, 'watcher');
 
+    async function clean(): Promise<void> {
+        if (await directoryExists(URI.toFsPath(baseURI))) {
+            (await fileService.delete(baseURI, { recursive: true }).unwrap());
+        }
+    }
+
     before(async () => {
         fileService = new FileService(new NullLogger());
         fileService.registerProvider(Schemas.FILE, new DiskFileSystemProvider());
+        await clean();
     });
 
-    after(async () => {
-        (await fileService.delete(baseURI, { recursive: true }).unwrap());
+    afterEach(async () => {
+        await clean();
     });
 
     test('watch file', async () => {
         const fileURI = URI.join(baseURI, 'file.txt');
         (await fileService.createFile(fileURI, DataBuffer.alloc(0), { overwrite: true }).unwrap());
         
-        const disposables: IDisposable[] = [];
-
         const watcher = new Watcher();
-        disposables.push(watcher.watch({
+        const cancel = await watcher.watch({
             resource: fileURI,
             recursive: false,
-        }));
+        });
 
-        const blocker = new Blocker<IRawResourceChangeEvents>();
-        disposables.push(watcher.onDidChange((e) => blocker.resolve(e)));
-
+        const blocker = new EventBlocker<IRawResourceChangeEvents>(watcher.onDidChange);
         (await fileService.writeFile(fileURI, DataBuffer.fromString('hello world'), { create: false, overwrite: true, }).unwrap());
-
         const e = await blocker.waiting();
         
         /**
@@ -52,38 +54,32 @@ suite('watcher-test', () => {
         // match check
         assert.ok(e.wrap().match(fileURI));
 
-        disposeAll(disposables);
+        cancel.dispose();
     });
 
-    test.skip('watch directory', async () => {
+    test('watch directory (recursive)', async function () {
         const subDirURI = URI.join(baseURI, 'subDir');
+        await fileService.createDir(subDirURI).unwrap();
+
         const fileURI = URI.join(subDirURI, 'file.txt');
 
-        const disposables: IDisposable[] = [];
-
         const watcher = new Watcher();
-        disposables.push(watcher.watch({
+        const cancel = await watcher.watch({
             resource: subDirURI,
-            recursive: true, // FIX: seems recursive is not working
-        }));
+            recursive: true,
+        });
 
-        const blocker = new Blocker<IRawResourceChangeEvents>();
-        disposables.push(watcher.onDidChange((e) => blocker.resolve(e)));
-
-
-        // await fileService.createDir(subDirURI);
-        (await fileService.writeFile(fileURI, DataBuffer.fromString('hello world'), { create: true, overwrite: true, }).unwrap());
-
-        const e = await blocker.waiting();
-        console.log(e);
-
-        assert.strictEqual(URI.parse(e.events[0]!.resource).scheme, Schemas.FILE);
-        assert.strictEqual(e.events[0]!.resource, URI.toString(fileURI));
+        const blocker = new EventBlocker<IRawResourceChangeEvents>(watcher.onDidChange);
+        await fileService.writeFile(fileURI, DataBuffer.fromString('hello world'), { create: true, overwrite: true, }).unwrap();
+        const events = await blocker.waiting();
         
-        // match check
-        assert.ok(e.wrap().match(fileURI));
+        assert.ok(events.anyAdded);
+        assert.ok(events.anyFile);
+        assert.ok(!events.anyDeleted);
+        assert.ok(!events.anyUpdated);
+        assert.ok(!events.anyDirectory);
+        assert.ok(events.wrap().match(fileURI));
 
-        disposeAll(disposables);
+        cancel.dispose();
     });
-
 });
